Increase audio element width to improve accessibility
Hello! 

I was browsing http://linenoise.io/sets/the_meadow/index.html today to listen to the tracklist. 

I noticed that the time slider of the <audio> element control panel is little small, thus making it hard to skip to any specific timestamp. 

Haven't worked much with CSS but searching around on StackOverflow, I found the proposal to set the width to 100%. I made the change and tried it out on mobile, tab and mac and it seem to work as expected.